feat(deployments): organize settings by groups (Issue #1438)
