+++ apps/ai-dial-admin/src/components/Containers/Fields/ContainerProperties.tsx
@@ -1,60 +1,35 @@
-import React, { FC, useCallback } from 'react';
+import React, { FC } from 'react';
 import classNames from 'classnames';
+
 import { Container } from '@/src/models/deployments/containers';
 import { ApplicationRoute } from '@/src/types/routes';
-import { EnvironmentVariable } from '@/src/models/deployments/variables';
-import BaseProperties from '@/src/components/Containers/Fields/BaseProperties';
+
+import BaseFields from '@/src/components/Containers/Fields/BaseFields/BaseFields';
 import ResourcesFields from '@/src/components/Containers/Fields/Resources/ResourcesFields';
-import PortField from '@/src/components/Common/PortField/PortField';
 import EnvVariables from '@/src/components/Containers/Fields/EnvVariables/EnvVariables';
-import ServingProperties from './ServingProperties';
-import Transport from '@/src/components/Containers/Fields/Transport/Transport';
+import EndpointConfiguration from '@/src/components/Containers/Fields/EndpointConfiguration/EndpointConfiguration';
 
 interface Props {
   container: Container;
   setContainer: (container: Container) => void;
+  names?: string[];
   isModal?: boolean;
   route: ApplicationRoute;
-  names?: string[];
 }
 
 const ContainerProperties: FC<Props> = ({ container, setContainer, isModal, route, names }) => {
-  const containerClassNames = classNames('flex flex-1 flex-col gap-4', !isModal && 'lg:w-[35%]');
-
-  const onVariablesChange = useCallback(
-    (variables: EnvironmentVariable[]) => {
-      setContainer({
-        ...container,
-        metadata: {
-          envs: variables,
-        },
-      });
-    },
-    [container, setContainer],
-  );
+  const containerClassNames = classNames('flex flex-1 flex-col gap-8', !isModal && 'lg:w-[35%]');
 
   return (
-    <div className={classNames('flex flex-col gap-4', !isModal && 'divide-y divide-primary')}>
+    <div className="flex flex-col gap-8">
       <div className={containerClassNames}>
-        <BaseProperties container={container} setContainer={setContainer} names={names} isModal={isModal} />
-        {!isModal && (
-          <>
-            {route === ApplicationRoute.ModelDeployments && (
-              <ServingProperties container={container} setContainer={setContainer} />
-            )}
-            <ResourcesFields container={container} setContainer={setContainer} route={route} />
-            {route === ApplicationRoute.McpDeployments && (
-              <div className="flex gap-4">
-                <Transport container={container} setContainer={setContainer} />
-              </div>
-            )}
-            {!isModal && <PortField route={route} container={container} setContainer={setContainer} />}
-          </>
-        )}
+        <BaseFields container={container} setContainer={setContainer} names={names} isModal={isModal} />
       </div>
       {!isModal && (
-        <div className="flex flex-col mt-8 pt-8">
-          <EnvVariables variables={container.metadata.envs || []} onChangeVariables={onVariablesChange} />
+        <div className="flex flex-col gap-y-8">
+          <EndpointConfiguration container={container} setContainer={setContainer} route={route} />
+          <EnvVariables container={container} setContainer={setContainer} />
+          <ResourcesFields container={container} setContainer={setContainer} route={route} />
         </div>
       )}
     </div>
